// src/app/models/contestantRanks.ts
export interface ContestantRanks {
    id: number;
    pointsCombined: number;
    rankCombined: number;
    pointsRiding: number;
    rankRiding: number;
    pointsRunning: number;
    rankRunning: number;
    pointsSwimming: number;
    rankSwimming: number;
    pointsWalking: number;
    rankWalking: number;
    position: number;
}